added new ldap-krb files and updated installation, config file and web2py_src folder

+++ baadalinstallation/installation.sh
@@ -1,10 +1,3 @@
-#####################################################################################
-#####################################################################################
-############################# OLD INSTALLATION SCRIPT ###############################
-#####################################################################################
-#####################################################################################
-
-
 ################################ FILE CONSTANTS USED ###########################
 
 Normal_pkg_lst=(ssh zip unzip tar openssh-server build-essential python2.7:python2.5 python-dev python-paramiko apache2 libapache2-mod-wsgi postfix wget libapache2-mod-gnutls  libvirt-bin apache2.2-common python-matplotlib python-reportlab mercurial python-libvirt sshpass)
@@ -23,7 +16,8 @@ Chk_Root_Login()
 	if test $username != "root"; then
 
   		echo "LOGIN AS SUPER USER(root) TO INSTALL BAADAL!!!"
-  		exit 1
+  		echo "EXITING INSTALLATION......................................"
+		exit 1
 	fi
 
 	echo "User Logged in as Root............................................"
@@ -32,40 +26,90 @@ Chk_Root_Login()
 #Function to place ldap and kerberos config files to there correct places
 Cnfgr_Ldap_Srvc()
 {
+	temp=0;
+
 	if test -f "/etc/krb5.conf";then
-		mv /etc/{krb5.conf,krb5.conf.bkp}
-	fi
-	cp -f ldap_krb/krb5.conf /etc/
-
+		if test -f "ldap-krb/krb5.conf";then
+			mv /etc/{krb5.conf,krb5.conf.bkp}
+			cp -f ldap-krb/krb5.conf /etc/
+		else
+			echo "ERROR: ldap-krb/krb5.conf"	
+	  		echo "EXITING INSTALLATION......................................"
+			exit 1
+		fi
+	fi 
+	
+	
 	if test -f "/etc/ldap.conf";then
-		mv /etc/{ldap.conf,ldap.conf.bkp}
+		if test -f "ldap-krb/ldap.conf";then
+			mv /etc/{ldap.conf,ldap.conf.bkp}
+			cp -f ldap-krb/ldap.conf /etc/
+		else
+			echo "ERROR: ldap-krb/ldap.conf"
+	  		echo "EXITING INSTALLATION......................................"
+			exit 1
+		fi
 	fi
-	cp -f ldap_krb/ldap.conf /etc/
-
+	
+	
 	if test -f "/etc/nsswitch.conf";then
-		mv /etc/{nsswitch.conf,nsswitch.conf.bkp}
+		if test -f "ldap-krb/nsswitch.conf";then
+			mv /etc/{nsswitch.conf,nsswitch.conf.bkp}
+			cp -f ldap-krb/nsswitch.conf /etc/
+		else
+			echo "ERROR: ldap-krb/nsswitch.conf"
+	  		echo "EXITING INSTALLATION......................................"
+			exit 1
+		fi	
 	fi
-	cp -f ldap_krb/nsswitch.conf /etc/
 
+	
 	if test -f "/etc/pam.d/common-account";then
-		mv /etc/pam.d/{common-account,common-account.bkp}
+		if test -f "ldap-krb/common-account";then
+			mv /etc/pam.d/{common-account,common-account.bkp}
+			cp -f ldap-krb/common-account /etc/pam.d/
+		else
+			echo "ERROR: ldap-krb/common-account"
+  			echo "EXITING INSTALLATION......................................"
+			exit 1
+		fi
 	fi
-	cp -f ldap_krb/common-account /etc/pam.d/
-
+	
+	
 	if test -f "/etc/pam.d/common-auth";then
-		mv /etc/pam.d/{common-auth,common-auth.bkp}
+		if test -f "ldap-krb/common-auth";then
+			mv /etc/pam.d/{common-auth,common-auth.bkp}
+			cp -f ldap-krb/common-auth /etc/pam.d/	
+		else
+			echo "ERROR: ldap-krb/common-auth"
+	  		echo "EXITING INSTALLATION......................................"
+			exit 1
+		fi
 	fi
-	cp -f ldap_krb/common-auth /etc/pam.d/
-
+	
+	
 	if test -f "/etc/pam.d/common-password";then
-		mv /etc/pam.d/{common-password,common-password.bkp}
+		if test -f "ldap-krb/common-password";then
+			mv /etc/pam.d/{common-password,common-password.bkp}
+			cp -f ldap-krb/common-password /etc/pam.d/	
+		else
+			echo "ERROR: ldap-krb/common-password"
+	  		echo "EXITING INSTALLATION......................................"
+			exit 1
+		fi
 	fi
-	cp -f ldap_krb/common-password /etc/pam.d/
-
+	
+	
 	if test -f "/etc/pam.d/common-session";then
-		mv /etc/pam.d/{common-session,common-session.bkp}
-	fi
-	cp -f ldap_krb/common-session /etc/pam.d/
+		if test -f "ldap-krb/common-session";then
+			mv /etc/pam.d/{common-session,common-session.bkp}
+			cp -f ldap-krb/common-session /etc/pam.d/		
+		else
+			echo "ERROR: ldap-krb/common-session"
+	  		echo "EXITING INSTALLATION......................................"
+			exit 1
+		fi
+	fi	
 }
 
 #Function to populate the list of packages to be installted
@@ -84,10 +128,10 @@ Populate_Pkg_Lst()
 	else
 		echo "Invalid Database Type!!!"
 		echo "Please Check Configuration File.........."
-		echo "Exiting.................................."
+  		echo "EXITING INSTALLATION......................................"
 		exit 1
-	fi			
 
+	fi			
 			
 	if [[ $authentication_type == "ldap" ]]; then
 	
@@ -101,7 +145,7 @@ Populate_Pkg_Lst()
 		
 		echo "Invalid Authentication Type!!!"
 		echo "Please Check Configuration File.........."
-		echo "Exiting.................................."
+  		echo "EXITING INSTALLATION......................................"
 		exit 1
 	fi
 
@@ -110,6 +154,10 @@ Populate_Pkg_Lst()
 #Function that install all the packages packages
 Instl_Pkgs()
 {	
+
+	echo "Updating System............."
+	apt-get update	
+
 	Pkg_lst=()
 	Populate_Pkg_Lst
 
@@ -125,49 +173,52 @@ Instl_Pkgs()
  	
 			if test $status -eq 1;  then 
 
-		        	echo "$pkg Package not installed................"
-				echo "Installing Package: $pkg.................."
+					echo "Installing Package: $pkg.................."
 				
-				if [[ "$pkg" =~ "mysql-server" ]]; then
+					if [[ "$pkg" =~ "mysql-server" ]]; then
 
-					echo "mysql-server-5.5 mysql-server/root_password password $mysql_password" | debconf-set-selections
-					echo "mysql-server-5.5 mysql-server/root_password_again password $mysql_password" | debconf-set-selections
-					DEBIAN_FRONTEND=noninteractive apt-get -y install $pkg --force-yes
-				else
-					DEBIAN_FRONTEND=noninteractive apt-get -y install $pkg --force-yes
-				fi
+						echo "mysql-server-5.5 mysql-server/root_password password $mysql_password" | debconf-set-selections
+						echo "mysql-server-5.5 mysql-server/root_password_again password $mysql_password" | debconf-set-selections
+						DEBIAN_FRONTEND=noninteractive apt-get -y install $pkg --force-yes
+					else
+						DEBIAN_FRONTEND=noninteractive apt-get -y install $pkg --force-yes
+					fi
 		        	
-				status=$?
+					status=$?
 		
-				if test $status -eq 0 ; then 
+					if test $status -eq 0 ; then 
 		      
-					echo "$pkg Package Installed Successfully" 
-					pkg_status=1
-					break
-			 	fi
+						echo "$pkg Package Installed Successfully" 
+						pkg_status=1
+						break
+					fi
 
-		        elif test $status -eq 0 ; then
+		     elif test $status -eq 0 ; then
 			
-				if [[ "$pkg" =~ "mysql-server" ]]; then
+#					if [[ "$pkg" =~ "mysql-server" ]]; then
 
-					echo "Mysql Already Installed!!!"
-					echo "Uninstall Mysql and Restart Installation Process................."
-					echo "Exiting Baadal Installation Process.............................."
-					exit
-				fi
+#						echo "Mysql Already Installed!!!"	
+#						echo "Uninstall Mysql and Restart Installation Process................."
+#						echo "Exiting Baadal Installation Process.............................."
+#						exit
+#					fi
 
 		        	echo "$pkg Package Already Installed" 
-				pkg_status=1;break     
-			fi
+					pkg_status=1;break     
+			 fi
 		done
+		
+		# end of FOR loop / package installation from pkg_multi_vrsn
 		
 		if test $pkg_status -eq 0; then
 			
 			echo "PACKAGE INSTALLATION UNSUCCESSFULL: ${pkg_multi_vrsn[@]} !!!"
 			echo "NETWORK CONNECTION ERROR/ REPOSITORY ERROR!!!"
-			exit 1 
+	  		echo "EXITING INSTALLATION......................................"
+			exit 1
 		fi	
 	done
+	# end of FOR loop / package installation from pkg_lst
 
 	echo "Packages Installed Successfully..................................."
 }
@@ -178,83 +229,61 @@ Setup_Web2py()
 {
 	Instl_Web2py=1
 	
-        if [ -d "/home/www-data/web2py/" ]; then
+    if [ -d "/home/www-data/web2py/" ]; then
 		echo "Web2py Already Exists!!!"
 		
 		while true; do
-			
 			case $reinstall_web2py in
-
-				n|N) Instl_Web2py=0; break; ;;
-			     y|Y|"") break; ;;
-				  *) echo "Invalid value of reinstall_web2py in config-file!!!!";break; ;;
+				n|N) 	Instl_Web2py=0; break; ;;
+				y|Y|"") break; ;;
+				*) 		echo "Invalid value of reinstall_web2py in config-file!!!!";break; ;;
 			esac
 		done
 	fi
 
 	if test $Instl_Web2py -ne 0; then
-		
-		wget http://www.web2py.com/examples/static/web2py_src.zip
-
-		if test ! -f "web2py_src.zip"; then
-
-			echo "UNABLE TO DOWNLOAD LATEST VERSION FROM REPOSITORY"
-
-			case $use_old_web2py_version in
-					
-			     y|Y|"") cp web2py_tmp.zip web2py_src.zip; break; ;;
-		  	        n|N) echo "Exiting web2py setup!!!";exit 1; ;;
-		       		  *) echo "Invalid value of use_old_web2py in config-file!!!";echo "Exiting the program";exit 1; ;;
-	    		esac
-	  
-			if [ ! -f "web2py_src.zip" ]; then
-				echo "some error in copying old version"
-				exit 1
-			fi
-		else
-			rm -rf web2py_tmp.zip; cp web2py_src.zip web2py_tmp.zip
-		fi
 
 		rm -rf web2py/
 		unzip web2py_src.zip
 
 		if test ! -d web2py/; then
-
 			echo "UNABLE TO EXTRACT WEB2PY!!!"
-			exit 1		
+	  		echo "EXITING INSTALLATION......................................"
+			exit 1
 		fi
-
-		rm -rf web2py_src.zip
+	
 		rm -rf /home/www-data/
 		mkdir /home/www-data/
-		cp -r web2py/ /home/www-data/web2py/
+		mv web2py/ /home/www-data/web2py/
 		
 		if test $? -ne '0'; then
 			echo "UNABLE TO SETUP WEB2PY!!!"
+	  		echo "EXITING INSTALLATION......................................"
 			exit 1
 		else
 			rm -rf web2py/
 		fi
-	fi		
+	fi
 
 	rm -rf /home/www-data/web2py/applications/baadal/
+	cp logging.conf /home/www-data/web2py/
 	cp -r baadal/ /home/www-data/web2py/applications/baadal/
-	cp config-file-baadal /home/www-data/web2py/applications/baadal/
 
 	if test $? -ne '0'; then
 		echo "UNABLE TO SETUP BAADAL!!!"
+  		echo "EXITING INSTALLATION......................................"
 		exit 1
 	fi
 
-	shopt -s nocasematch
-	case $authentication_type in
-
-		ldap) mv /home/www-data/web2py/applications/baadal/models/functions.py.ldap /home/www-data/web2py/applications/baadal/models/functions.py
-		      rm /home/www-data/web2py/applications/baadal/models/functions.py.localdb
-		      ;;
-	     localdb) mv /home/www-data/web2py/applications/baadal/models/functions.py.localdb /home/www-data/web2py/applications/baadal/models/functions.py
-		      rm /home/www-data/web2py/applications/baadal/models/functions.py.ldap
-	esac
+#	shopt -s nocasematch
+#	case $authentication_type in
+#
+#		ldap) mv /home/www-data/web2py/applications/baadal/models/functions.py.ldap /home/www-data/web2py/applications/baadal/models/functions.py
+#		      rm /home/www-data/web2py/applications/baadal/models/functions.py.localdb
+#		      ;;
+#	     localdb) mv /home/www-data/web2py/applications/baadal/models/functions.py.localdb /home/www-data/web2py/applications/baadal/models/functions.py
+#		      rm /home/www-data/web2py/applications/baadal/models/functions.py.ldap
+#	esac
 
 	chown -R www-data:www-data /home/www-data/
 
@@ -278,50 +307,38 @@ Enbl_Modules()
 		mysql) /etc/init.d/mysql restart
 
 			    if test $? -ne 0; then
-				echo "UNABLE TO RESTART MYSQL!!!"
-				exit 1
+					echo "UNABLE TO RESTART MYSQL!!!"
+			  		echo "EXITING INSTALLATION......................................"
+					exit 1
 			    fi
 
-			    if [ -d /var/lib/mysql/baadal ] ; then
+			    if [ -d /var/lib/mysql/$mysql_database_name ] ; then
+			    	echo "$mysql_database_name already exists!!!"
+			    	echo "Please remove the $mysql_database_name database and restart the installation process..."
+			  		echo "EXITING INSTALLATION......................................"
+					exit 1
+			    fi
 
-				echo "Dropping Previous Database Baadal........."
-				
-				rm -rf /var/lib/mysql/baadal
-			
-				if test $? -ne 0; then
-			
-					echo "ERROR IN REMOVING EXISTING BAADAL DATABASE!!!"
-					echo "Please remove it manaully and start the installation"
-					echo "Exiting Installation Process.."	
-					exit 1			    
-				fi	
-			     fi
-
-			     echo "Creating Database Baadal......................"
+			    echo "Creating Database $mysql_database_name......................"
 
 
-			     while true; do
-			
-				mysql -uroot -p$mysql_password -e "create database baadal" 2> temp
+				mysql -uroot -p$mysql_password -e "create database $mysql_database_name" 2> temp
 
-				is_valid_paswd=`grep "ERROR 1045 (28000): Access denied for user 'root'@'localhost' " temp | wc -l`
+				if test $? -ne 0;then
+					cat temp					
+					is_valid_paswd=`grep "ERROR 1045 (28000): Access denied for user 'root'@'localhost' " temp | wc -l`
+					rm -rf temp	
 
-				rm -rf temp
+					if test $is_valid_paswd -ne 0; then
+						echo "INVALID MYSQL PASSWORD!!!!"				    
+					else
+						break
+					fi
 
-				if test $is_valid_paswd -ne 0; then
-			
-				    echo "INVALID MYSQL PASSWORD!!!!"				    
-				else
-
-				    break
-				fi
-			     done
-
-			     if test $? -ne 0; then
-
-				echo "UNABLE TO CREATE DATABASE!!!"
-				exit 1
-			     fi
+					echo "UNABLE TO CREATE DATABASE!!!"
+			  		echo "EXITING INSTALLATION......................................"
+					exit 1						
+				fi					
 	esac
 
 }
@@ -333,7 +350,7 @@ Create_SSL_Certi()
 	echo "creating Self Signed Certificate................................."
 	openssl genrsa 1024 > /etc/apache2/ssl/self_signed.key
 	chmod 400 /etc/apache2/ssl/self_signed.key
-	openssl req -new -x509 -nodes -sha1 -days 365 -key /etc/apache2/ssl/self_signed.key -config config-file-baadal > /etc/apache2/ssl/self_signed.cert
+	openssl req -new -x509 -nodes -sha1 -days 365 -key /etc/apache2/ssl/self_signed.key -config config-file > /etc/apache2/ssl/self_signed.cert
 	openssl x509 -noout -fingerprint -text < /etc/apache2/ssl/self_signed.cert > /etc/apache2/ssl/self_signed.info
 }
 
@@ -421,13 +438,34 @@ Rewrite_Apache_Conf()
 		  ErrorLog /var/log/apache2/error.log
 		</VirtualHost>
 		' > /etc/apache2/sites-available/default
-		
+
+#	echo "Enabling rewrite-mod to rewrite http request to https..........."
+
+#	a2enmod rewrite
+	
 	echo "Restarting Apache................................................"
 	
 	/etc/init.d/apache2 restart
 
+#	echo "writing rule to redirect http requests to https.................."		
+
+#	echo '
+#		<VirtualHost *:80>
+#		  RewriteEngine On
+#		  RewriteCond %{SERVER_PORT} ^80$
+#		  RewriteRule .* https://%{SERVER_NAME}%{REQUEST_URI} [R,L]
+#		</VirtualHost>
+#	' >> /etc/apache2/sites-enabled/web2py
+
+#	echo "Restarting Apache................................................"
+	
+#	/etc/init.d/apache2 restart
+
+
 	if test $? -ne 0; then
 		echo "UNABLE TO RESTART APACHE!!!"
+		echo "CHECK APACHE LOGS FOR DETAILS!!!"
+		echo "EXITING INSTALLATION......................................"
 		exit 1
 	fi
 	
@@ -437,17 +475,14 @@ Rewrite_Apache_Conf()
 Start_Web2py()
 {
 	if test ! -d "/var/www/"; then
-
 		echo "PROBLEM IN APACHE!!!"
+  		echo "EXITING INSTALLATION......................................"
 		exit 1
 
 	elif test ! -f "/var/www/.ssh/id_rsa"; then
-	
 		if test ! -d "/var/www/.ssh"; then
-
 			mkdir /var/www/.ssh
 			chown -R www-data:www-data /var/www/.
-
 		fi
 
 		su www-data -c "ssh-keygen -t rsa -f /var/www/.ssh/id_rsa -N \"\""
@@ -457,27 +492,28 @@ Start_Web2py()
 		echo "PUBLIC KEY ALREADY EXISTS!!!"
 	fi
 
+	echo "setting up web2py.................."
 	cd /home/www-data/web2py
 	sudo -u www-data python -c "from gluon.widget import console; console();"
 	sudo -u www-data python -c "from gluon.main import save_password; save_password(\"$web2py_password\",443)"
 
 	cd applications/baadal/
 
-	nohup python -u /home/www-data/web2py/web2py.py -S baadal -M -R /home/www-data/web2py/applications/baadal/private/hostpower.py >> hostpower.out &
-	nohup python -u /home/www-data/web2py/web2py.py -S baadal -M -R /home/www-data/web2py/applications/baadal/private/process.py >> process.out &
+#	nohup python -u /home/www-data/web2py/web2py.py -S baadal -M -R /home/www-data/web2py/applications/baadal/private/hostpower.py &
+#	nohup python -u /home/www-data/web2py/web2py.py -S baadal -M -R /home/www-data/web2py/applications/baadal/private/process.py >> process.out &
 	
-	val=`cat /etc/crontab | grep "\*\/45 \* \* \* \* www-data python /home/www-data/web2py/applications/baadal/cron/check_baadal_status.py" | wc -l`
+#	val=`cat /etc/crontab | grep "\*\/45 \* \* \* \* www-data python /home/www-data/web2py/applications/baadal/cron/check_baadal_status.py" | wc -l`
 
-	if [[ $val -eq 0 ]]; then
+#	if [[ $val -eq 0 ]]; then
+#		echo "*/45 * * * * www-data python /home/www-data/web2py/applications/baadal/cron/check_baadal_status.py" >> /etc/crontab
+#	fi
 
-		echo "*/45 * * * * www-data python /home/www-data/web2py/applications/baadal/cron/check_baadal_status.py" >> /etc/crontab
-	fi
 }
 
 ################################ SCRIPT ########################################
 
 #Including Config File to the current script
-. ./config-file-baadal 2>>/dev/null
+. ./config-file 2>>/dev/null
 
 Chk_Root_Login
 Instl_Pkgs
@@ -487,6 +523,7 @@ Create_SSL_Certi
 Rewrite_Apache_Conf
 Start_Web2py
 echo "done!"
+
 
 #####################################################################################
 #####################################################################################
